Move project splitting/unifying code to separate library
This allows to reuse parts of the code in LanguageForge S/R without
requiring any UI.

- move UI independent parts of the code in a new library
  LibFLEXBridge-ChorusPlugin
- move corresponding unit tests to LibFLExBridge-ChorusPluginTests
- some cleanup
- changed visibility of some methods from internal to public
- adjust to new version of Chorus that uses MEF
- make use of MEF to get file handler types
  (in FieldWorksCommonFileHandler.cs)

Note that the library has to be called *-ChorusPlugin.dll so that it
gets picked up when Chorus scans for assemblies. This is necessary
because the library implements FieldWorksCommonFileHandler which
derives from IChorusFileTypeHandler. This naming is necessary
indepent of Chorus using MEF.